feat: limit calls to get_light_level to once per 10 minutes
Refactored most of the code to simplify the interface and to
reduce the amount of unnecessary processing required by the
consumer.

BREAKING CHANGE: class names have changed as follows:

  Connect => SomaConnect
  Shade   => SomaShade

The values returned by the SomaShade methods have beenn simplified
to return either a boolean to denote whether the action succeeded
or not or the value of the requested property.

Signed-off-by: Avi Miller <me@dje.li>